WARNING: plenty of rogue code roams these classes, tried to fix that but my computer has thrown two blue-screens in their defense, so I'm leaving the rogue code alone.
Added a new way of setting up a parking spot, with tags. This will represent a real life situation (with a robot) much more accurately.
The car has 5 new sensors, each with an id (from 1000 to 1010) and the floor also has five tags (matching with the sensor id's). Each sensor will read it's own tag only.
The fitness function is now calculated using the distance from each sensors' center to its given tag center. The lower the fitness value, the better.
There's still an issue, however, in the possibility of real life implementation, and in order to use this fitness function, we'd have to use some wireless comunications between the robot and the tags in order to evaluate the distance between both. Maybe we could use time as a fitness value, multiplied by the sum of boolean values indicating if each of the tags is being detected or not.
Even though I believe the second option's use is much more correct in the real world, the first is probably much more efficient.